refactor: LottoNumber, LottoTicket, Lotto class
